requirements.txt
paho-mqtt>=1.6
python-dotenv>=1.0
pandas>=2.2
matplotlib>=3.8
pytest>=8.0
